+++ backend/src/main/java/com/example/demo/exception/AccountNotFoundException.java
@@ -1,5 +1,5 @@
 package com.example.demo.exception;
 
 public class AccountNotFoundException extends RuntimeException {
-    public AccountNotFoundException(int id) { super("Account ID " + id + " not found."); }
+    public AccountNotFoundException(Long id) { super("Account ID " + id + " not found."); }
 }
